refactor tests and comments
