+++ packages/sources/asseto-finance/test/integration/fixtures/fixtures.adapter.ts
@@ -46,41 +46,6 @@ export const mockReserveResponseSuccess = (fundId: number): nock.Scope =>
     )
     .persist()
 
-export const mockReserveResponseRipcord = (fundId: number): nock.Scope =>
-  nock('https://open.syncnav.com/api', {
-    encodedQueryParams: true,
-  })
-    .get(`/funds/${fundId}/reserves`)
-    .reply(
-      200,
-      () => ({
-        code: 0,
-        message: 'success',
-        data: {
-          fundId: 8,
-          fundName: 'CashPlus_BSC',
-          totalAum: '800',
-          totalSupply: '9534.885',
-          updatedAt: '2025-09-22 20:00:01',
-          ripcord: true,
-          ripcordDetails: 'some ripcord details',
-        },
-        timestamp: 1758542882,
-        traceID: 'c872cb2e6f9967180ebbdc38d4087023',
-      }),
-      [
-        'Content-Type',
-        'application/json',
-        'Connection',
-        'close',
-        'Vary',
-        'Accept-Encoding',
-        'Vary',
-        'Origin',
-      ],
-    )
-    .persist()
-
 export const mockNavResponseSuccess = (fundId: number): nock.Scope =>
   nock('https://open.syncnav.com/api/funds', {
     encodedQueryParams: true,
@@ -127,53 +92,4 @@ export const mockNavResponseSuccess = (fundId: number): nock.Scope =>
         'Origin',
       ],
     )
-    .persist()
-
-export const mockAuthResponseUnauthorized = (): nock.Scope =>
-  nock('https://open.syncnav.com')
-    .post('/api/oauth/token')
-    .reply(401, {
-      error: 'invalid_client',
-      error_description: 'Client authentication failed',
-    })
-    .persist()
-
-export const mockAuthResponseServerError = (): nock.Scope =>
-  nock('https://open.syncnav.com')
-    .post('/api/oauth/token')
-    .reply(500, {
-      error: 'server_error',
-      error_description: 'Internal server error',
-    })
-    .persist()
-
-export const mockAuthResponseMissingToken = (): nock.Scope =>
-  nock('https://open.syncnav.com')
-    .post('/api/oauth/token')
-    .reply(200, {
-      token_type: 'Bearer',
-      scopes: [],
-      expires: 604800,
-    })
-    .persist()
-
-export const mockAuthResponseMissingExpiry = (): nock.Scope =>
-  nock('https://open.syncnav.com')
-    .post('/api/oauth/token')
-    .reply(200, {
-      access_token: 'some_token',
-      token_type: 'Bearer',
-      scopes: [],
-    })
-    .persist()
-
-export const mockAuthResponseInvalidExpiry = (): nock.Scope =>
-  nock('https://open.syncnav.com')
-    .post('/api/oauth/token')
-    .reply(200, {
-      access_token: 'some_token',
-      token_type: 'Bearer',
-      scopes: [],
-      expires: 0,
-    })
     .persist()
